Imported WebKit commit 356d83016b090995d08ad568f2d2c243aa55e831 (http://svn.webkit.org/repository/webkit/trunk@126147)
New snapshot including various build fixes for newer Qt 5

+++ Source/WebCore/platform/graphics/chromium/cc/CCLayerTreeHostImpl.cpp
@@ -24,29 +24,31 @@
 
 #include "config.h"
 
-#include "cc/CCLayerTreeHostImpl.h"
+#include "CCLayerTreeHostImpl.h"
 
+#include "CCActiveGestureAnimation.h"
+#include "CCDamageTracker.h"
+#include "CCDebugRectHistory.h"
+#include "CCDelayBasedTimeSource.h"
+#include "CCFontAtlas.h"
+#include "CCFrameRateCounter.h"
+#include "CCHeadsUpDisplayLayerImpl.h"
+#include "CCLayerIterator.h"
+#include "CCLayerTreeHost.h"
+#include "CCLayerTreeHostCommon.h"
+#include "CCMathUtil.h"
+#include "CCOverdrawMetrics.h"
+#include "CCPageScaleAnimation.h"
+#include "CCPrioritizedTextureManager.h"
+#include "CCRenderPassDrawQuad.h"
+#include "CCRenderingStats.h"
+#include "CCScrollbarAnimationController.h"
+#include "CCScrollbarLayerImpl.h"
+#include "CCSettings.h"
+#include "CCSingleThreadProxy.h"
 #include "LayerRendererChromium.h"
 #include "TextStream.h"
 #include "TraceEvent.h"
-#include "cc/CCActiveGestureAnimation.h"
-#include "cc/CCDamageTracker.h"
-#include "cc/CCDebugRectHistory.h"
-#include "cc/CCDelayBasedTimeSource.h"
-#include "cc/CCFontAtlas.h"
-#include "cc/CCFrameRateCounter.h"
-#include "cc/CCLayerIterator.h"
-#include "cc/CCLayerTreeHost.h"
-#include "cc/CCLayerTreeHostCommon.h"
-#include "cc/CCOverdrawMetrics.h"
-#include "cc/CCPageScaleAnimation.h"
-#include "cc/CCPrioritizedTextureManager.h"
-#include "cc/CCRenderPassDrawQuad.h"
-#include "cc/CCRenderingStats.h"
-#include "cc/CCScrollbarAnimationController.h"
-#include "cc/CCScrollbarLayerImpl.h"
-#include "cc/CCSettings.h"
-#include "cc/CCSingleThreadProxy.h"
 #include <wtf/CurrentTime.h>
 
 using WebKit::WebTransformationMatrix;
@@ -118,11 +120,13 @@ CCLayerTreeHostImpl::CCLayerTreeHostImpl(const CCLayerTreeSettings& settings, CC
     , m_sourceFrameNumber(-1)
     , m_rootScrollLayerImpl(0)
     , m_currentlyScrollingLayerImpl(0)
+    , m_hudLayerImpl(0)
     , m_scrollingLayerIdFromPreviousTree(-1)
+    , m_scrollDeltaIsInScreenSpace(false)
     , m_settings(settings)
     , m_deviceScaleFactor(1)
     , m_visible(true)
-    , m_contentsTexturesWerePurgedSinceLastCommit(false)
+    , m_contentsTexturesPurged(false)
     , m_memoryAllocationLimitBytes(CCPrioritizedTextureManager::defaultMemoryAllocationLimit())
     , m_pageScale(1)
     , m_pageScaleDelta(1)
@@ -159,7 +163,6 @@ void CCLayerTreeHostImpl::commitComplete()
     // Recompute max scroll position; must be after layer content bounds are
     // updated.
     updateMaxScrollPosition();
-    m_contentsTexturesWerePurgedSinceLastCommit = false;
 }
 
 bool CCLayerTreeHostImpl::canDraw()
@@ -176,7 +179,7 @@ bool CCLayerTreeHostImpl::canDraw()
         TRACE_EVENT_INSTANT0("cc", "CCLayerTreeHostImpl::canDraw no layerRenderer");
         return false;
     }
-    if (m_contentsTexturesWerePurgedSinceLastCommit) {
+    if (m_contentsTexturesPurged) {
         TRACE_EVENT_INSTANT0("cc", "CCLayerTreeHostImpl::canDraw contents textures purged");
         return false;
     }
@@ -252,6 +255,7 @@ void CCLayerTreeHostImpl::calculateRenderSurfaceLayerList(CCLayerList& renderSur
 {
     ASSERT(renderSurfaceLayerList.isEmpty());
     ASSERT(m_rootLayerImpl);
+    ASSERT(m_layerRenderer); // For maxTextureSize.
 
     {
         TRACE_EVENT0("cc", "CCLayerTreeHostImpl::calcDrawEtc");
@@ -512,8 +516,10 @@ bool CCLayerTreeHostImpl::prepareToDraw(FrameData& frame)
 
 void CCLayerTreeHostImpl::releaseContentsTextures()
 {
+    if (m_contentsTexturesPurged)
+        return;
     m_resourceProvider->deleteOwnedResources(CCRenderer::ContentPool);
-    m_contentsTexturesWerePurgedSinceLastCommit = true;
+    m_contentsTexturesPurged = true;
     m_client->setNeedsCommitOnImplThread();
 }
 
@@ -543,10 +549,15 @@ void CCLayerTreeHostImpl::drawLayers(const FrameData& frame)
     // RenderWidget.
     m_fpsCounter->markBeginningOfFrame(currentTime());
 
-    m_layerRenderer->drawFrame(frame.renderPasses, frame.renderPassesById);
-
     if (m_settings.showDebugRects())
         m_debugRectHistory->saveDebugRectsForCurrentFrame(m_rootLayerImpl.get(), *frame.renderSurfaceLayerList, frame.occludingScreenSpaceRects, settings());
+
+    // Because the contents of the HUD depend on everything else in the frame, the contents
+    // of its texture are updated as the last thing before the frame is drawn.
+    if (m_hudLayerImpl)
+        m_hudLayerImpl->updateHudTexture(m_resourceProvider.get());
+
+    m_layerRenderer->drawFrame(frame.renderPasses, frame.renderPassesById);
 
     // Once a RenderPass has been drawn, its damage should be cleared in
     // case the RenderPass will be reused next frame.
@@ -829,6 +840,8 @@ bool CCLayerTreeHostImpl::ensureRenderSurfaceLayerList()
 {
     if (!m_rootLayerImpl)
         return false;
+    if (!m_layerRenderer)
+        return false;
 
     // We need both a non-empty render surface layer list and a root render
     // surface to be able to iterate over the visible layers.
@@ -884,36 +897,95 @@ CCInputHandlerClient::ScrollStatus CCLayerTreeHostImpl::scrollBegin(const IntPoi
 
     if (potentiallyScrollingLayerImpl) {
         m_currentlyScrollingLayerImpl = potentiallyScrollingLayerImpl;
+        // Gesture events need to be transformed from screen coordinates to local layer coordinates
+        // so that the scrolling contents exactly follow the user's finger. In contrast, wheel
+        // events are already in local layer coordinates so we can just apply them directly.
+        m_scrollDeltaIsInScreenSpace = (type == Gesture);
         return ScrollStarted;
     }
     return ScrollIgnored;
 }
 
-void CCLayerTreeHostImpl::scrollBy(const IntSize& scrollDelta)
+static FloatSize scrollLayerWithScreenSpaceDelta(CCLayerImpl& layerImpl, const FloatPoint& screenSpacePoint, const FloatSize& screenSpaceDelta)
+{
+    // Layers with non-invertible screen space transforms should not have passed the scroll hit
+    // test in the first place.
+    ASSERT(layerImpl.screenSpaceTransform().isInvertible());
+    WebTransformationMatrix inverseScreenSpaceTransform = layerImpl.screenSpaceTransform().inverse();
+
+    // First project the scroll start and end points to local layer space to find the scroll delta
+    // in layer coordinates.
+    bool startClipped, endClipped;
+    FloatPoint screenSpaceEndPoint = screenSpacePoint + screenSpaceDelta;
+    FloatPoint localStartPoint = CCMathUtil::projectPoint(inverseScreenSpaceTransform, screenSpacePoint, startClipped);
+    FloatPoint localEndPoint = CCMathUtil::projectPoint(inverseScreenSpaceTransform, screenSpaceEndPoint, endClipped);
+
+    // In general scroll point coordinates should not get clipped.
+    ASSERT(!startClipped);
+    ASSERT(!endClipped);
+    if (startClipped || endClipped)
+        return FloatSize();
+
+    // Apply the scroll delta.
+    FloatSize previousDelta(layerImpl.scrollDelta());
+    layerImpl.scrollBy(localEndPoint - localStartPoint);
+
+    // Calculate the applied scroll delta in screen space coordinates.
+    FloatPoint actualLocalEndPoint = localStartPoint + layerImpl.scrollDelta() - previousDelta;
+    FloatPoint actualScreenSpaceEndPoint = CCMathUtil::mapPoint(layerImpl.screenSpaceTransform(), actualLocalEndPoint, endClipped);
+    ASSERT(!endClipped);
+    if (endClipped)
+        return FloatSize();
+    return actualScreenSpaceEndPoint - screenSpacePoint;
+}
+
+static FloatSize scrollLayerWithLocalDelta(CCLayerImpl& layerImpl, const FloatSize& localDelta)
+{
+    FloatSize previousDelta(layerImpl.scrollDelta());
+    layerImpl.scrollBy(localDelta);
+    return layerImpl.scrollDelta() - previousDelta;
+}
+
+void CCLayerTreeHostImpl::scrollBy(const IntPoint& viewportPoint, const IntSize& scrollDelta)
 {
     TRACE_EVENT0("cc", "CCLayerTreeHostImpl::scrollBy");
     if (!m_currentlyScrollingLayerImpl)
         return;
 
     FloatSize pendingDelta(scrollDelta);
-    pendingDelta.scale(1 / m_pageScaleDelta);
-
-    for (CCLayerImpl* layerImpl = m_currentlyScrollingLayerImpl; layerImpl && !pendingDelta.isZero(); layerImpl = layerImpl->parent()) {
+    for (CCLayerImpl* layerImpl = m_currentlyScrollingLayerImpl; layerImpl; layerImpl = layerImpl->parent()) {
         if (!layerImpl->scrollable())
             continue;
-        FloatSize previousDelta(layerImpl->scrollDelta());
-        layerImpl->scrollBy(pendingDelta);
-        // Reset the pending scroll delta to zero if the layer was able to move along the requested
-        // axis. This is to ensure it is possible to scroll exactly to the beginning or end of a
-        // scroll area regardless of the scroll step. For diagonal scrolls this also avoids applying
-        // the scroll on one axis to multiple layers.
-        if (previousDelta.width() != layerImpl->scrollDelta().width())
-            pendingDelta.setWidth(0);
-        if (previousDelta.height() != layerImpl->scrollDelta().height())
-            pendingDelta.setHeight(0);
+
+        FloatSize appliedDelta;
+        if (m_scrollDeltaIsInScreenSpace)
+            appliedDelta = scrollLayerWithScreenSpaceDelta(*layerImpl, viewportPoint, pendingDelta);
+        else
+            appliedDelta = scrollLayerWithLocalDelta(*layerImpl, pendingDelta);
+
+        // If the layer wasn't able to move, try the next one in the hierarchy.
+        float moveThresholdSquared = 0.1 * 0.1;
+        if (appliedDelta.diagonalLengthSquared() < moveThresholdSquared)
+            continue;
+
+        // If the applied delta is within 45 degrees of the input delta, bail out to make it easier
+        // to scroll just one layer in one direction without affecting any of its parents.
+        float angleThreshold = 45;
+        if (CCMathUtil::smallestAngleBetweenVectors(appliedDelta, pendingDelta) < angleThreshold) {
+            pendingDelta = FloatSize();
+            break;
+        }
+
+        // Allow further movement only on an axis perpendicular to the direction in which the layer
+        // moved.
+        FloatSize perpendicularAxis(-appliedDelta.height(), appliedDelta.width());
+        pendingDelta = CCMathUtil::projectVector(pendingDelta, perpendicularAxis);
+
+        if (flooredIntSize(pendingDelta).isZero())
+            break;
     }
 
-    if (!scrollDelta.isZero() && pendingDelta.isEmpty()) {
+    if (!scrollDelta.isZero() && flooredIntSize(pendingDelta).isEmpty()) {
         m_client->setNeedsCommitOnImplThread();
         m_client->setNeedsRedrawOnImplThread();
     }
